rename: 페이지 구성 컴포넌트 파일명 변경 (#2)
